# app/marksman/internal/server/swagger/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /v1/metric/strategy/{strategyUID}:
        get:
            tags:
                - StrategyMetric
            operationId: StrategyMetric_GetStrategyMetric
            parameters:
                - name: strategyUID
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/marksman.api.v1.StrategyMetricItem'
        post:
            tags:
                - StrategyMetric
            operationId: StrategyMetric_SaveStrategyMetric
            parameters:
                - name: strategyUID
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/marksman.api.v1.SaveStrategyMetricRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/marksman.api.v1.SaveStrategyMetricReply'
...
    /v1/strategies:
        get:
            tags:
                - Strategy
            operationId: Strategy_ListStrategy
            parameters:
                - name: keyword
                  in: query
                  schema:
                    type: string
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: status
                  in: query
                  schema:
                    type: integer
                    format: enum
                - name: strategyGroupUID
                  in: query
                  schema:
                    type: string
                - name: type
                  in: query
                  schema:
                    type: integer
                    format: enum
                - name: driver
                  in: query
                  schema:
                    type: integer
                    format: enum
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/marksman.api.v1.ListStrategyReply'
    /v1/strategies/select:
        get:
            tags:
                - Strategy
            operationId: Strategy_SelectStrategy
            parameters:
                - name: keyword
                  in: query
                  schema:
                    type: string
                - name: limit
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: lastUid
                  in: query
                  schema:
                    type: string
                - name: status
                  in: query
                  schema:
                    type: integer
                    format: enum
                - name: strategyGroupUids
                  in: query
                  description: When set, only strategies belonging to these strategy groups are returned (for alert page filter).
                  schema:
                    type: array
                    items:
                        type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/marksman.api.v1.SelectStrategyReply'
    /v1/strategy:
        post:
            tags:
                - Strategy
            operationId: Strategy_CreateStrategy
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/marksman.api.v1.CreateStrategyRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/marksman.api.v1.CreateStrategyReply'
    /v1/strategy-group:
        post:
            tags:
                - Strategy
            operationId: Strategy_CreateStrategyGroup
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/marksman.api.v1.CreateStrategyGroupRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/marksman.api.v1.CreateStrategyGroupReply'
    /v1/strategy-group/{uid}:
        get:
            tags:
                - Strategy
            operationId: Strategy_GetStrategyGroup
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/marksman.api.v1.StrategyGroupItem'
        put:
            tags:
                - Strategy
            operationId: Strategy_UpdateStrategyGroup
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/marksman.api.v1.UpdateStrategyGroupRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/marksman.api.v1.UpdateStrategyGroupReply'
        delete:
            tags:
                - Strategy
            operationId: Strategy_DeleteStrategyGroup
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/marksman.api.v1.DeleteStrategyGroupReply'
    /v1/strategy-group/{uid}/status:
        put:
            tags:
                - Strategy
            operationId: Strategy_UpdateStrategyGroupStatus
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/marksman.api.v1.UpdateStrategyGroupStatusRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/marksman.api.v1.UpdateStrategyGroupStatusReply'
    /v1/strategy-groups:
        get:
            tags:
                - Strategy
            operationId: Strategy_ListStrategyGroup
            parameters:
                - name: keyword
                  in: query
                  schema:
                    type: string
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: status
                  in: query
                  schema:
                    type: integer
                    format: enum
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/marksman.api.v1.ListStrategyGroupReply'
    /v1/strategy-groups/select:
        get:
            tags:
                - Strategy
            operationId: Strategy_SelectStrategyGroup
            parameters:
                - name: keyword
                  in: query
                  schema:
                    type: string
                - name: limit
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: lastUID
                  in: query
                  schema:
                    type: string
                - name: status
                  in: query
                  schema:
                    type: integer
                    format: enum
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/marksman.api.v1.SelectStrategyGroupReply'
...
components:
    schemas:
...
        marksman.api.v1.CreateStrategyGroupReply:
            type: object
            properties:
                uid:
                    type: string
...
        marksman.api.v1.CreateStrategyReply:
            type: object
            properties:
                uid:
                    type: string
...
        marksman.api.v1.ListStrategyReply:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/marksman.api.v1.StrategyItem'
                total:
                    type: string
                page:
                    type: integer
                    format: int32
                pageSize:
                    type: integer
                    format: int32
...
        marksman.api.v1.SaveStrategyMetricReply:
            type: object
            properties: {}
        marksman.api.v1.SaveStrategyMetricRequest:
            type: object
            properties:
                strategyUID:
                    type: string
                expr:
                    type: string
                labels:
                    type: object
                    additionalProperties:
                        type: string
                datasourceUIDs:
                    type: array
                    items:
                        type: string
                summary:
                    type: string
                description:
                    type: string
...
        marksman.api.v1.StrategyMetricItem:
            type: object
            properties:
                strategyUID:
                    type: string
                expr:
                    type: string
                labels:
                    type: object
                    additionalProperties:
                        type: string
                summary:
                    type: string
                description:
                    type: string
                datasourceUIDs:
                    type: array
                    items:
                        type: string
                levels:
                    type: array
                    items:
                        $ref: '#/components/schemas/marksman.api.v1.StrategyMetricLevelItem'
                createdAt:
                    type: string
                updatedAt:
                    type: string
                strategy:
                    $ref: '#/components/schemas/marksman.api.v1.StrategyItem'
...