# app/marksman/pkg/api/v1/strategy_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v5.29.2
// source: marksman/api/v1/strategy.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationStrategyCreateStrategy = "/marksman.api.v1.Strategy/CreateStrategy"
const OperationStrategyCreateStrategyGroup = "/marksman.api.v1.Strategy/CreateStrategyGroup"
const OperationStrategyDeleteStrategy = "/marksman.api.v1.Strategy/DeleteStrategy"
const OperationStrategyDeleteStrategyGroup = "/marksman.api.v1.Strategy/DeleteStrategyGroup"
const OperationStrategyGetStrategy = "/marksman.api.v1.Strategy/GetStrategy"
const OperationStrategyGetStrategyGroup = "/marksman.api.v1.Strategy/GetStrategyGroup"
const OperationStrategyListStrategy = "/marksman.api.v1.Strategy/ListStrategy"
const OperationStrategyListStrategyGroup = "/marksman.api.v1.Strategy/ListStrategyGroup"
const OperationStrategySelectStrategy = "/marksman.api.v1.Strategy/SelectStrategy"
const OperationStrategySelectStrategyGroup = "/marksman.api.v1.Strategy/SelectStrategyGroup"
const OperationStrategyUpdateStrategy = "/marksman.api.v1.Strategy/UpdateStrategy"
const OperationStrategyUpdateStrategyGroup = "/marksman.api.v1.Strategy/UpdateStrategyGroup"
const OperationStrategyUpdateStrategyGroupStatus = "/marksman.api.v1.Strategy/UpdateStrategyGroupStatus"
const OperationStrategyUpdateStrategyStatus = "/marksman.api.v1.Strategy/UpdateStrategyStatus"

type StrategyHTTPServer interface {
	CreateStrategy(context.Context, *CreateStrategyRequest) (*CreateStrategyReply, error)
	CreateStrategyGroup(context.Context, *CreateStrategyGroupRequest) (*CreateStrategyGroupReply, error)
	DeleteStrategy(context.Context, *DeleteStrategyRequest) (*DeleteStrategyReply, error)
	DeleteStrategyGroup(context.Context, *DeleteStrategyGroupRequest) (*DeleteStrategyGroupReply, error)
	GetStrategy(context.Context, *GetStrategyRequest) (*StrategyItem, error)
	GetStrategyGroup(context.Context, *GetStrategyGroupRequest) (*StrategyGroupItem, error)
	ListStrategy(context.Context, *ListStrategyRequest) (*ListStrategyReply, error)
	ListStrategyGroup(context.Context, *ListStrategyGroupRequest) (*ListStrategyGroupReply, error)
	SelectStrategy(context.Context, *SelectStrategyRequest) (*SelectStrategyReply, error)
	SelectStrategyGroup(context.Context, *SelectStrategyGroupRequest) (*SelectStrategyGroupReply, error)
	UpdateStrategy(context.Context, *UpdateStrategyRequest) (*UpdateStrategyReply, error)
	UpdateStrategyGroup(context.Context, *UpdateStrategyGroupRequest) (*UpdateStrategyGroupReply, error)
	UpdateStrategyGroupStatus(context.Context, *UpdateStrategyGroupStatusRequest) (*UpdateStrategyGroupStatusReply, error)
	UpdateStrategyStatus(context.Context, *UpdateStrategyStatusRequest) (*UpdateStrategyStatusReply, error)
}

func RegisterStrategyHTTPServer(s *http.Server, srv StrategyHTTPServer) {
	r := s.Route("/")
	r.POST("/v1/strategy-group", _Strategy_CreateStrategyGroup0_HTTP_Handler(srv))
	r.PUT("/v1/strategy-group/{uid}", _Strategy_UpdateStrategyGroup0_HTTP_Handler(srv))
	r.PUT("/v1/strategy-group/{uid}/status", _Strategy_UpdateStrategyGroupStatus0_HTTP_Handler(srv))
	r.DELETE("/v1/strategy-group/{uid}", _Strategy_DeleteStrategyGroup0_HTTP_Handler(srv))
	r.GET("/v1/strategy-group/{uid}", _Strategy_GetStrategyGroup0_HTTP_Handler(srv))
	r.GET("/v1/strategy-groups", _Strategy_ListStrategyGroup0_HTTP_Handler(srv))
	r.GET("/v1/strategy-groups/select", _Strategy_SelectStrategyGroup0_HTTP_Handler(srv))
	r.POST("/v1/strategy", _Strategy_CreateStrategy0_HTTP_Handler(srv))
	r.PUT("/v1/strategy/{uid}", _Strategy_UpdateStrategy0_HTTP_Handler(srv))
	r.PUT("/v1/strategy/{uid}/status", _Strategy_UpdateStrategyStatus0_HTTP_Handler(srv))
	r.DELETE("/v1/strategy/{uid}", _Strategy_DeleteStrategy0_HTTP_Handler(srv))
	r.GET("/v1/strategy/{uid}", _Strategy_GetStrategy0_HTTP_Handler(srv))
	r.GET("/v1/strategies", _Strategy_ListStrategy0_HTTP_Handler(srv))
	r.GET("/v1/strategies/select", _Strategy_SelectStrategy0_HTTP_Handler(srv))
}

func _Strategy_CreateStrategyGroup0_HTTP_Handler(srv StrategyHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateStrategyGroupRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationStrategyCreateStrategyGroup)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateStrategyGroup(ctx, req.(*CreateStrategyGroupRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CreateStrategyGroupReply)
		return ctx.Result(200, reply)
	}
}

func _Strategy_UpdateStrategyGroup0_HTTP_Handler(srv StrategyHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateStrategyGroupRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationStrategyUpdateStrategyGroup)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateStrategyGroup(ctx, req.(*UpdateStrategyGroupRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateStrategyGroupReply)
		return ctx.Result(200, reply)
	}
}

func _Strategy_UpdateStrategyGroupStatus0_HTTP_Handler(srv StrategyHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateStrategyGroupStatusRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationStrategyUpdateStrategyGroupStatus)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateStrategyGroupStatus(ctx, req.(*UpdateStrategyGroupStatusRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateStrategyGroupStatusReply)
		return ctx.Result(200, reply)
	}
}

func _Strategy_DeleteStrategyGroup0_HTTP_Handler(srv StrategyHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DeleteStrategyGroupRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationStrategyDeleteStrategyGroup)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DeleteStrategyGroup(ctx, req.(*DeleteStrategyGroupRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*DeleteStrategyGroupReply)
		return ctx.Result(200, reply)
	}
}

func _Strategy_GetStrategyGroup0_HTTP_Handler(srv StrategyHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetStrategyGroupRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationStrategyGetStrategyGroup)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetStrategyGroup(ctx, req.(*GetStrategyGroupRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*StrategyGroupItem)
		return ctx.Result(200, reply)
	}
}

func _Strategy_ListStrategyGroup0_HTTP_Handler(srv StrategyHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListStrategyGroupRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationStrategyListStrategyGroup)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListStrategyGroup(ctx, req.(*ListStrategyGroupRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListStrategyGroupReply)
		return ctx.Result(200, reply)
	}
}

func _Strategy_SelectStrategyGroup0_HTTP_Handler(srv StrategyHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in SelectStrategyGroupRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationStrategySelectStrategyGroup)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.SelectStrategyGroup(ctx, req.(*SelectStrategyGroupRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*SelectStrategyGroupReply)
		return ctx.Result(200, reply)
	}
}

func _Strategy_CreateStrategy0_HTTP_Handler(srv StrategyHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateStrategyRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationStrategyCreateStrategy)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateStrategy(ctx, req.(*CreateStrategyRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CreateStrategyReply)
		return ctx.Result(200, reply)
	}
}

func _Strategy_UpdateStrategy0_HTTP_Handler(srv StrategyHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateStrategyRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationStrategyUpdateStrategy)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateStrategy(ctx, req.(*UpdateStrategyRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateStrategyReply)
		return ctx.Result(200, reply)
	}
}

func _Strategy_UpdateStrategyStatus0_HTTP_Handler(srv StrategyHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateStrategyStatusRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationStrategyUpdateStrategyStatus)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateStrategyStatus(ctx, req.(*UpdateStrategyStatusRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateStrategyStatusReply)
		return ctx.Result(200, reply)
	}
}

func _Strategy_DeleteStrategy0_HTTP_Handler(srv StrategyHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DeleteStrategyRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationStrategyDeleteStrategy)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DeleteStrategy(ctx, req.(*DeleteStrategyRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*DeleteStrategyReply)
		return ctx.Result(200, reply)
	}
}

func _Strategy_GetStrategy0_HTTP_Handler(srv StrategyHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetStrategyRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationStrategyGetStrategy)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetStrategy(ctx, req.(*GetStrategyRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*StrategyItem)
		return ctx.Result(200, reply)
	}
}

func _Strategy_ListStrategy0_HTTP_Handler(srv StrategyHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListStrategyRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationStrategyListStrategy)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListStrategy(ctx, req.(*ListStrategyRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListStrategyReply)
		return ctx.Result(200, reply)
	}
}

func _Strategy_SelectStrategy0_HTTP_Handler(srv StrategyHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in SelectStrategyRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationStrategySelectStrategy)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.SelectStrategy(ctx, req.(*SelectStrategyRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*SelectStrategyReply)
		return ctx.Result(200, reply)
	}
}

type StrategyHTTPClient interface {
	CreateStrategy(ctx context.Context, req *CreateStrategyRequest, opts ...http.CallOption) (rsp *CreateStrategyReply, err error)
	CreateStrategyGroup(ctx context.Context, req *CreateStrategyGroupRequest, opts ...http.CallOption) (rsp *CreateStrategyGroupReply, err error)
	DeleteStrategy(ctx context.Context, req *DeleteStrategyRequest, opts ...http.CallOption) (rsp *DeleteStrategyReply, err error)
	DeleteStrategyGroup(ctx context.Context, req *DeleteStrategyGroupRequest, opts ...http.CallOption) (rsp *DeleteStrategyGroupReply, err error)
	GetStrategy(ctx context.Context, req *GetStrategyRequest, opts ...http.CallOption) (rsp *StrategyItem, err error)
	GetStrategyGroup(ctx context.Context, req *GetStrategyGroupRequest, opts ...http.CallOption) (rsp *StrategyGroupItem, err error)
	ListStrategy(ctx context.Context, req *ListStrategyRequest, opts ...http.CallOption) (rsp *ListStrategyReply, err error)
	ListStrategyGroup(ctx context.Context, req *ListStrategyGroupRequest, opts ...http.CallOption) (rsp *ListStrategyGroupReply, err error)
	SelectStrategy(ctx context.Context, req *SelectStrategyRequest, opts ...http.CallOption) (rsp *SelectStrategyReply, err error)
	SelectStrategyGroup(ctx context.Context, req *SelectStrategyGroupRequest, opts ...http.CallOption) (rsp *SelectStrategyGroupReply, err error)
	UpdateStrategy(ctx context.Context, req *UpdateStrategyRequest, opts ...http.CallOption) (rsp *UpdateStrategyReply, err error)
	UpdateStrategyGroup(ctx context.Context, req *UpdateStrategyGroupRequest, opts ...http.CallOption) (rsp *UpdateStrategyGroupReply, err error)
	UpdateStrategyGroupStatus(ctx context.Context, req *UpdateStrategyGroupStatusRequest, opts ...http.CallOption) (rsp *UpdateStrategyGroupStatusReply, err error)
	UpdateStrategyStatus(ctx context.Context, req *UpdateStrategyStatusRequest, opts ...http.CallOption) (rsp *UpdateStrategyStatusReply, err error)
}

type StrategyHTTPClientImpl struct {
	cc *http.Client
}

func NewStrategyHTTPClient(client *http.Client) StrategyHTTPClient {
	return &StrategyHTTPClientImpl{client}
}

func (c *StrategyHTTPClientImpl) CreateStrategy(ctx context.Context, in *CreateStrategyRequest, opts ...http.CallOption) (*CreateStrategyReply, error) {
	var out CreateStrategyReply
	pattern := "/v1/strategy"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationStrategyCreateStrategy))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *StrategyHTTPClientImpl) CreateStrategyGroup(ctx context.Context, in *CreateStrategyGroupRequest, opts ...http.CallOption) (*CreateStrategyGroupReply, error) {
	var out CreateStrategyGroupReply
	pattern := "/v1/strategy-group"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationStrategyCreateStrategyGroup))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *StrategyHTTPClientImpl) DeleteStrategy(ctx context.Context, in *DeleteStrategyRequest, opts ...http.CallOption) (*DeleteStrategyReply, error) {
	var out DeleteStrategyReply
	pattern := "/v1/strategy/{uid}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationStrategyDeleteStrategy))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *StrategyHTTPClientImpl) DeleteStrategyGroup(ctx context.Context, in *DeleteStrategyGroupRequest, opts ...http.CallOption) (*DeleteStrategyGroupReply, error) {
	var out DeleteStrategyGroupReply
	pattern := "/v1/strategy-group/{uid}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationStrategyDeleteStrategyGroup))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *StrategyHTTPClientImpl) GetStrategy(ctx context.Context, in *GetStrategyRequest, opts ...http.CallOption) (*StrategyItem, error) {
	var out StrategyItem
	pattern := "/v1/strategy/{uid}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationStrategyGetStrategy))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *StrategyHTTPClientImpl) GetStrategyGroup(ctx context.Context, in *GetStrategyGroupRequest, opts ...http.CallOption) (*StrategyGroupItem, error) {
	var out StrategyGroupItem
	pattern := "/v1/strategy-group/{uid}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationStrategyGetStrategyGroup))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *StrategyHTTPClientImpl) ListStrategy(ctx context.Context, in *ListStrategyRequest, opts ...http.CallOption) (*ListStrategyReply, error) {
	var out ListStrategyReply
	pattern := "/v1/strategies"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationStrategyListStrategy))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *StrategyHTTPClientImpl) ListStrategyGroup(ctx context.Context, in *ListStrategyGroupRequest, opts ...http.CallOption) (*ListStrategyGroupReply, error) {
	var out ListStrategyGroupReply
	pattern := "/v1/strategy-groups"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationStrategyListStrategyGroup))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *StrategyHTTPClientImpl) SelectStrategy(ctx context.Context, in *SelectStrategyRequest, opts ...http.CallOption) (*SelectStrategyReply, error) {
	var out SelectStrategyReply
	pattern := "/v1/strategies/select"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationStrategySelectStrategy))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *StrategyHTTPClientImpl) SelectStrategyGroup(ctx context.Context, in *SelectStrategyGroupRequest, opts ...http.CallOption) (*SelectStrategyGroupReply, error) {
	var out SelectStrategyGroupReply
	pattern := "/v1/strategy-groups/select"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationStrategySelectStrategyGroup))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *StrategyHTTPClientImpl) UpdateStrategy(ctx context.Context, in *UpdateStrategyRequest, opts ...http.CallOption) (*UpdateStrategyReply, error) {
	var out UpdateStrategyReply
	pattern := "/v1/strategy/{uid}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationStrategyUpdateStrategy))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *StrategyHTTPClientImpl) UpdateStrategyGroup(ctx context.Context, in *UpdateStrategyGroupRequest, opts ...http.CallOption) (*UpdateStrategyGroupReply, error) {
	var out UpdateStrategyGroupReply
	pattern := "/v1/strategy-group/{uid}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationStrategyUpdateStrategyGroup))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *StrategyHTTPClientImpl) UpdateStrategyGroupStatus(ctx context.Context, in *UpdateStrategyGroupStatusRequest, opts ...http.CallOption) (*UpdateStrategyGroupStatusReply, error) {
	var out UpdateStrategyGroupStatusReply
	pattern := "/v1/strategy-group/{uid}/status"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationStrategyUpdateStrategyGroupStatus))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *StrategyHTTPClientImpl) UpdateStrategyStatus(ctx context.Context, in *UpdateStrategyStatusRequest, opts ...http.CallOption) (*UpdateStrategyStatusReply, error) {
	var out UpdateStrategyStatusReply
	pattern := "/v1/strategy/{uid}/status"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationStrategyUpdateStrategyStatus))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}